refactor: retire TS CLI; Python CLI is now the only armoriq login
- src/cli/ deleted: login/logout/whoami commands. The Python SDK
  now owns the CLI surface (armoriq login via OAuth 2.0 device-code
  flow).
- src/credentials.ts kept (moved up from src/cli/): the JS SDK
  still reads ~/.armoriq/credentials.json written by armoriq login.
- package.json: drop bin entry and CLI-only deps (commander, open).
- client.ts: error message points users at pip install armoriq-sdk
  && armoriq login instead of npx.
- _build_env.ts: staging IAP now uses the iap-staging.armoriq.ai
  custom domain (previously the long Cloud Run auto URL).
